1. Layout change for auth routes
