+++ src/app/(authroutes)/layout.tsx
@@ -1,0 +1,51 @@
+import "../globals.css";
+import type { Metadata } from "next";
+import { Inter } from "next/font/google";
+import { ToggleProvider } from "@/contextProviders/ToggleProvider";
+import NextAuthProvider from "@/contextProviders/sessionProvider";
+import { Toaster } from "react-hot-toast";
+import { Providers } from "@/store/provider";
+import { Suspense } from "react";
+
+const inter = Inter({ subsets: ["latin"] });
+
+export const metadata: Metadata = {
+  title: "TPC Portal",
+  description: "Generated by create next app",
+};
+
+interface Props {
+  children: React.ReactNode;
+}
+
+const AuthLayout = async ({ children }: Props) => {
+  const className = inter.className;
+
+  return (
+    <html lang="en">
+      <body className={className}>
+        <ToggleProvider>
+          <div className="flex flex-col h-screen">
+            <Toaster />
+
+            <NextAuthProvider>
+              <div className="flex-auto flex h-[92vh] ">
+                {/* sidebar and main content share this space */}
+                <Suspense fallback={<>Loading...</>}></Suspense>
+                <Providers>
+                  <div className="w-screen min-h-screen h-max bg-[url(/images/iiti_bg.JPG)] bg-cover bg-fixed bg-center bg-no-repeat absolute">
+                    <div className="w-full min-h-full h-max bg-gradient-to-b from-black/40 to-black/90 bg-opacity-50 absolute">
+                      {children}
+                    </div>
+                  </div>
+                </Providers>
+              </div>
+            </NextAuthProvider>
+          </div>
+        </ToggleProvider>
+      </body>
+    </html>
+  );
+};
+
+export default AuthLayout;
