+++ src/app/(authroutes)/recruiter/signup/page.tsx
@@ -2,11 +2,7 @@ import RecruiterSignup from "@/components/loginForms/recruiterSignup";
 import React from "react";
 
 const Signup = () => {
-  return (
-    <>
-      <RecruiterSignup />
-    </>
-  );
+  return <RecruiterSignup />;
 };
 
 export default Signup;
